dev: Move `arg` to `cmd` package
