+++ cmd/ttn-lw-migrate/main.go
@@ -21,7 +21,7 @@ import (
 	_ "go.thethings.network/lorawan-stack-migrate/pkg/source/ttnv2"      // TTNv2 source
 	_ "go.thethings.network/lorawan-stack-migrate/pkg/source/ttnv3"      // TTS source
 
-	_ "go.thethings.network/lorawan-stack-migrate/pkg/arg" // Init after initializing sources
+	// _ "go.thethings.network/lorawan-stack-migrate/pkg/arg" // Init after initializing sources
 
 	"go.thethings.network/lorawan-stack-migrate/cmd"
 )
